무한 참조 문제 해결 중

+++ src/main/java/com/example/meetup_study/user/fakeUser/FakeUserServiceImpl.java
@@ -2,6 +2,8 @@ package com.example.meetup_study.user.fakeUser;
 
 import com.example.meetup_study.user.domain.User;
 import com.example.meetup_study.user.domain.dto.ResponseUserDto;
+import com.example.meetup_study.user.fakeUser.exception.UserInvalidRequestException;
+import com.example.meetup_study.user.fakeUser.exception.UserNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -17,10 +19,18 @@ public class FakeUserServiceImpl {
     private final FakeRepository fakeUserRepository;
 
     public Optional<User> createFakeUser(User user){
+
+        if(fakeUserRepository.existsByEmail(user.getEmail())){
+            throw new UserInvalidRequestException("중복 유저 회원가입 요청 실패입니다.");
+        }
+
         return Optional.ofNullable(fakeUserRepository.save(user));
     }
 
     public Optional<User> deleteFakeUser(User user) {
+        if(!fakeUserRepository.existsByEmail(user.getEmail())){
+            throw new UserNotFoundException();
+        }
         fakeUserRepository.delete(user);
         return Optional.ofNullable(user);
     }
